% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/localgeo-package.R
\docType{package}
\name{localgeo}
\alias{localgeo}
\alias{localgeo-package}
\title{A package to geocode US City & State pairs}
\description{
A package to geocode US City & State pairs
}
\author{
Bob Rudis (@hrbrmstr)
}

